Level sisteme giriş yapıldı

+++ Assets/Scripts/Data/ValueObjects/PoolData.cs
@@ -3,6 +3,6 @@ using System;
 [Serializable]
 public struct PoolData
 {
-    public int ReqviredObjectCount;
+    public byte ReqviredObjectCount;
 
 }
